Add copy/move to register/kill-ring
In addition to adding the ability to move and copy to the kill-ring
and specified register, the copy and move commands were refactored to
share logic much of their logic.

Debug messages have been removed, so `sam` is less noisy now.

Fixed a bug, where `sam` could not be used non-interactively.